Move DICOM header recording functions to dcmd
